[Pure Storage] Fix issue with NVMe LACP ports not being identified
When using NVMe as a dataplane, target ports are identified
directly from the backend.

When an LACP bond is created this was not being correctly
identified as valid.

This patch resolves this issue.

This patch is standalone for 2024.1 as the codebase changed
for 2024.2 and this bug is addressed in a seperate patch for
that release.

Closes-Bug: #2101859
Change-Id: Ic511d040fec57fb681f09c4f7459b222e16b84f7
(cherry picked from commit d4c63178975939f7df99a33a66c6704565d06e4c)