[Pure Storage] Fix QoS setting for cloned volumes
Volumes created from glance images on the source FlashArray using
volume cloning are not being assigned the correct QoS specifications
associated with the requested volume type.

This patch resolves that issue.

This patch is standalone for 2024.1 as the codebase changed
for 2024.2 and this bug is addressed in a seperate patch for
that release.

Partial-Bug: #2100547
Change-Id: Ic3bc26eee419ecae7377f05e9488b809cd3b843a
(cherry picked from commit 7b571e5f54216820d74cfd0609a3b2b625bef027)