Handle volume_clear_* opts consistently
There are 3 volume_clear_* options for lvm thick volumes, but only 2
of them are currently read from a backend-specific section of the
cinder conf file.  So an operator could define all 3 for a particular
backend, but one of them would still use the setting from the DEFAULT
section of the config file, which would be difficult to troubleshoot.
It would be better if all 3 behaved the same way.

Also enhance one of the unit tests related to the volume_clear feature.

Closes-bug: #2106680
Change-Id: I623ca7f9061b3e3edb9660da65bd9d64309f52d4