[Pure Storage] Add capacity based backend QoS options
Add new QoS spec parameters based on the capacity of the volume.

New parameters are maxIOPS_per_GB and maxBWS_per_GB.

If provied in the same QoS spec as the original max values for
IOPS or BWS, the max values will take precedence.

Capacity based QoS settings will be calculated and applied
for the following actions:

- create volume
- create volume from snapshot
- extend volume
- clone volume
- manage volume
- retype volume

QoS settings are not applied to volumes within volume groups
as QoS for these is controlled at the volume group level.

Implements: blueprint pure-per-gb-qos

Change-Id: I79b5180ac48140f747761d0c04f7bd62d9995b43