[Pure Storage] Disable management of secure tenant volumes and snaps
FlashArray has enabled secure multi-tenancy, but these volumes and
associated snapshots, although visible to the driver, are not eligible
for management.

This patch ensures that these volumes and snapshots are not listed in
the manageable volumes or snapshots list and reports an error if one
is still selected to be managed.

Change-Id: I85e65ebffecf029c9e29e6ba1d01dfacd28efff7