[Pure Storage] Consistent hostnames across replicated backends
Hostname objects on FlashArray backends were being created using a
construct that used a random UUID. This meant that on replicated
backends, the same host was being assigned different hostname
objects on the backend.

This fix replaces the random UUID with the hypervisor system UUID,
if available, to allow the hostname object to be consistent
across the replicated backends.

Change-Id: I7d883a9a1fb4dfbd92f38e11530aa3a0b7c42c77
Signed-off-by: Simon Dodsley <simon@purestorage.com>