NetApp: iSCSI/FC detach operation fails when multiple initiators connected
During the detach operation of iSCSI and Fibre Channel (FC) sessions,
the process fails when multiple initiators are connected to LUNS.
Investigation reveals that the root cause is related to improper
handling of initiator identifiers due to the use of a space character
as a delimiter.

Specifically, the NetApp's REST API  expects initiator IDs to be parsed
and processed as distinct entities. However, when multiple initiators
are listed or passed as a single string separated by spaces, the parsing
logic incorrectly splits or merges these identifiers. This results in
failure to correctly identify and detach each initiator session, causing
the overall detach operation to fail.

To fix the above, updated the delimeter from space character to comma.

Closes-Bug: #2114993
Change-Id: Ia143e729a3e84e22fcecf6f5bac8b0b3ae311553
Signed-off-by: Saikumar Pulluri <saikumar1016@gmail.com>