NFS driver: Fix fail to resize NFS volume with snapshots
The NFS driver may fail to update the virtual size of the
Image-volume to the new size when it has snapshots created, due
to image_utils incorrectly rejecting qcow2 backing files as they
were not allowed.

This patch explicitly tells `image_utils.qemu_img_info` to allow
the use of qcow2 backing files when performing a extend operation.

Closes-Bug: #2103742
Change-Id: I0582daac259dbda78481b56378562350def59598