NFS driver: Set volume format for snapshot operations
On snapshot operations, the NFS driver does not properly update
the volume format in admin_metadata to correspond to the format
of the active snaphot, as in the case of offline snapshot
creation, in which the driver does update the format to match
the newly created qcow2 snapshot, resulting in the resizing
operation to fail due to the format mismatch in the
`qemu-img resize` command.

This patch fixes such behavior by updating the volume format
in volume_admin_metadata whenever the active snapshot is replaced.

Change-Id: I7cacdd9b78820d272058e9c35f926e23b4c87ba4
Signed-off-by: Fernando Ferraz Silva <fesilva@redhat.com>