NFS driver: Use database format in initialize_connection
The `qemu_img_info` operation in initialize_connection may not be
able to determine whether the volume format is `raw` or `qcow2`,
when an inner qcow2 image is writen into a volume. This issue
commonly happens when Glance is configured with Cinder/NFS as
image store.

This patch explicitily tells `qemu_img_info` operation to use
the database format stored in admin_metadata `format` field instead
of trying a autodiscover, to ensure volume metadata such as the
virtual_size will be gathered correctly.

Partial-Bug: #2073146
Change-Id: I28df6ddfa4f62b440e01918958352ab186234f87
Signed-off-by: Fernando Ferraz Silva <fesilva@redhat.com>