NFS driver: Fix driver fails to verify virtual size for Glance with Cinder/NFS
Glance images stored in Cinder with NFS as backend are sized in
the Cinder database to match the data size of the Image-volume
instead of the virtual_size. This prevents storage
resource to be unnecessarily allocated, but may also results into
errors in operations that requires checking for volume size
consistency.

Such checks aims to prevent users from extending NFS volumes
from within the virtual machine and outside of Cinder scope, by
comparing the virtual size of the NFS image with its size stored
in Cinder's database.

This patch changes how volume size is verified to fail only when
a volume virtual size has been extended in size instead of trying
to verify the exact size in bytes.

Closes-Bug: #2073146
Change-Id: I5c2216b72dda6bb3316ef30b352a6a8efe3a0429
Signed-off-by: Fernando Ferraz Silva <fesilva@redhat.com>