[NetApp-ZAPI] Enabling snapshot creation for flexgroup pool
Currently, we don't have provision to create a snapshot for flexgroup pool
through ZAPI, with this change we will have the support for the same.
So after these changes, this is how it looks. For a FlexGroup pool, the
ZAPI operation uses the NFS generic driver. When it comes to REST, if the
ONTAP version is below 9.14, the operation depends on the NFS generic
driver. However, for ONTAP versions 9.14 and above, it relies on the ONTAP
file clone API.

Closes-Bug: #2119644

Change-Id: Iec73a2fd95716e30f18315759328423076d7c824
Signed-off-by: Saikumar Pulluri <saikumar1016@gmail.com>