Fix: optimize create volume for NFS
When we enable optimized path for creating a bootable volume from
image, we call the driver's ``create_cloned_volume`` method to
clone the image-volume to create the new volume.

When Glance is using Cinder as a backend and Cinder is backed by
generic NFS, we encounter the following error:

AttributeError: 'Volume' object has no attribute 'obj_context'

After looking into the optimization code path, we can see that the
``create_cloned_volume`` method is called with SQLAlchemy object[1][2]
instead of the oslo versioned volume object. When the remotefs
driver tries to fetch the context from it (obj_context), we fail.
This is not faced by the LVM or RBD driver as they only access
the properties that are available in the SQLAlchemy object.

The right thing to do here is to pass the OVO object to the drivers
which this patch intends to do.
NOTE: This issue is not specific to the NFS driver but since that
operation is broken, the releasenote is intentionally written that
way to aware operators that the issue is fixed.

[1] https://opendev.org/openstack/cinder/src/commit/2a6c3565438554fd857ddbb57bb65ca05a17d91b/cinder/volume/flows/manager/create_volume.py#L750-L755
[2] https://opendev.org/openstack/cinder/src/commit/2a6c3565438554fd857ddbb57bb65ca05a17d91b/cinder/volume/flows/manager/create_volume.py#L781

Closes-Bug: #2120933
Change-Id: Ibfbcff01c761a569587e453c952ae10bbee30fbb
Signed-off-by: Rajat Dhasmana <rajatdhasmana@gmail.com>