Delete image cache volumes when they're unusable
Delete a cached image volume when it cannot be cloned again due to
a snapshot limit. Previously, the volume was evicted from the cache
but the volume itself wasn't deleted. This left behind a volume
that consumed quota but was otherwise unusable.

Very few drivers have a snapshot limit, but those that do may raise
SnapshotLimitReached when asked to clone a volume in the image volume
cache. When this occurs, the volume is evicted from the cache and a
new image volume is created in order to seed a fresh cache entry.

It was previously thought the evicted volume couldn't be deleted due
to a dependency chain on all of its clones. However, this assumption
is false because the image volume cache's design is predicated on the
ability to delete these volumes.
- Source volumes in the cache may be deleted to enforce the cache's
  max_count or max_size settings.
- Source volumes in the cache will be deleted if the image data is
  no longer valid.

Now, when an image cache volume becomes unusable because it cannot
be cloned, it's deleted and not just evicted from the cache.

Closes-Bug: #2112403
Signed-off-by: Alan Bishop <abishop@redhat.com>
Change-Id: I01ff7b0be5da3657ab455582e34c24c5de8e2cf2