Add option to restrict arbitrary containers for backups
This patch provides implementation of new configuration option
which aims to allow operators to restrict arbitrary container
creation by users for backends based off chunkeddriver.

While one approach could be to avoid executing put_container overall,
and place all logic in chunkeddriver, this patch goes a different
route and implements option support in each driver, as this allows
to process and test raised exceptions in more driver-oriented way.
This also allows to fail a bit more early, on the contrary to failure
when data attempts to be placed in non-existing location.

Implements-blueprint: restrict-backup-container-creation
Change-Id: I0afe33b4fdfa0f5d3dbda394ae817f14eea67c03
Signed-off-by: Dmitriy Rabotyagov <dmitriy.rabotyagov@cleura.com>