Remove mypy-wrap.sh
This is not needed any more (and arguably never was: sorry about that).
All relevant configuration can be defined in pyproject.toml instead. To
do this, we enable a few strict options and add a note regarding
enabling a few more.

We also remove mypy from test-requirements.txt: this isn't needed for
unit/functional testing.

Change-Id: I7dec75c242db15eac19b74ca1518955d6118301f
Signed-off-by: Stephen Finucane <stephenfin@redhat.com>