Fix clone across pools with image volume cache
The clone across pool feature provides optimization
to create a volume in one pool by cloning from an
image volume cache volume residing in another pool.
However, the current logic doesn't work as we try
to fetch the ``clone_across_pools`` capability
while initializing the volume driver but the
property is populated when the backend reports
capabilities which happens after the initialization
has completed resulting in ``clone_across_pools``
value always being passed as ``False`` to the
Image volume cache class.
This patch corrects the behavior by fetching the
``clone_across_pools`` capability during volume
creation time from the cached capabilities so
that we get the correct value as well as doesn't
do an "expensive" call to the backend to fetch
capabilities again.

This patch also changes the logging level from
exception to debug when falling back to the
generic mechanism if cloning fails.
The exception trace is misleading (as seen in
LP:2127220) since it is not a real failure
but just an indication of optimization not
performed.

Closes-Bug: #2127675
Closes-Bug: #2127220

Change-Id: I503d2ba7656f151ef99dc18c9fd7a605466c2de5
Signed-off-by: Rajat Dhasmana <rajatdhasmana@gmail.com>