image_utils: Detect missing device before calling qemu-img convert
We have observed failures related to device management where
this code path is called and the device in /dev/ does not exist.

In this situation, qemu-img convert will create the dest file
in /dev/ and write data to it. Ensure that this does not happen
and the operation fails instead.

Closes-Bug: #2132083
Change-Id: Ide448f3809a4169a757578310066a5c3e0dffdae
Signed-off-by: Eric Harney <eharney@redhat.com>