stable: Remove some non-voting jobs from stable checks
This proposes removing some of the non-voting jobs from the stable check
pipeline. The non-voting jobs are rarely checked (voting ones barely
are), and presumably these have already been validated in the upstream
branch. So to reduce noise and CI time, this removes some of the less
relevant jobs.

If there is agreement, this should then be cherry picked to other open
stable branches as well.

Change-Id: I73550df24e76cf5490c91af27a9bc855d4633ce7
Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>