[Pure Storage] Fix performance stats collection error
In arrays that are very new and/or with very little or no
workload, the existing call to gather the last 30 seconds of
performance telemetry to send to cinder scheduler is returning
an empty dictionary.

This patch changes the collection parameters to increase the
chance of getting a correct response, and also stops the
failure should any empty response still be obtained.

Closes-Bug: #2123855
Change-Id: I21ee5118eacba9c39b8b637dc54f5a6a4abc7b55
Signed-off-by: Simon Dodsley <simon@purestorage.com>